Disable font sanitizing for `cargo build`
The fontsan crate wraps a C++ library and requires cmake. That means it
doesn't build with `cargo build` on most Windows systems. It was easy
to comment out the code to get the build to continue. I assume the
sanitizing was for added for security reasons so it would probably be a
good idea to write a pure Rust sanitizer that could be used instead.